+++ src/pages/Products/Products.tsx
@@ -58,19 +58,20 @@ export const Products = () => {
     }
 
     const handleDelete = (e: SyntheticEvent) => {
-
         const delItem = e.currentTarget.id;
         //@ts-ignore
         const nameDelItem = e.currentTarget.dataset.name
-        deleteProduct(delItem)
+        if(window.confirm(`Czy jesteś pewien, ze chcesz usunąć ${nameDelItem}`)) {
+            deleteProduct(delItem)
+            setMess(`Usunięto produkt ${nameDelItem} z bazy`)
+            setSuccess(false)
+            const newData = [...product].filter(item => item.id !== delItem);
+            setProduct(newData);
+            closeNotification()
+        } else {
+            return
+        }
 
-        setMess(`Usunięto produkt ${nameDelItem} z bazy`)
-        setSuccess(false)
-
-        const newData = [...product].filter(item => item.id !== delItem);
-
-        setProduct(newData);
-        closeNotification()
     }
     // todo przy błedach usuwania itp postaraj się wyświetlać belkę z informacją
